clean up and a little testing
